Implement reshape via backpermute to simplify fusion
Fusion currently assumes that a buffer variable is always used with the same shape. If the program contains reshapes, this is not true. We should thus either make fusion more complex, or handle reshapes differently. We chose to do the latter. A reshape becomes a backpermute. If the backpermute is fused, this has no overhead. We may later on add a pass to remove these backpermutes *after* fusion, and make the cost function of fusion aware that it is not beneficial to remove the allocation of backpermute, since there is no actual allocation.